+++ app/components/timeseries/timeseries-service.js
@@ -186,6 +186,7 @@ angular.module('timeseries')
         unit: '',
         color: '', // Defined on asset.timeseries
         order: '', // Defined on asset.timeseries
+        valueType: '',
         labels: {
           x: '',
           y: ''
@@ -198,6 +199,7 @@ angular.module('timeseries')
         var graphTimeseries = angular.copy(graphTimeseriesTemplate);
         graphTimeseries.data = ts.events;
         graphTimeseries.id = ts.uuid;
+        graphTimeseries.valueType = ts.value_type;
         graphTimeseries = addColorAndOrderAndUnit(graphTimeseries);
         result.push(graphTimeseries);
       });